requirements.txt
Flask==2.2.3
jsonschema==4.17.3
rfc3339-validator==0.1.4
pyjwt==2.6.0
flask_cors==3.0.10
python-dotenv==1.0.0
click~=8.1.3
pip~=23.1
attrs~=23.1.0
MarkupSafe~=2.1.2
Werkzeug~=2.2.3
Jinja2~=3.1.2
itsdangerous~=2.1.2
six~=1.16.0
pyrsistent~=0.19.3
setuptools~=67.7.1
zipp~=3.15.0